Set exclusions in the multiruby dexelas material
This sets the non-excluded hkls to be the first 5 for all materials
except ruby, where it is 8. The reason the first 8 hkls are included
for ruby is because the `active_hkls` in the config file point to
the first 6 hkls, where hkls with a structure factor less than 5% are
skipped (which skips 2 hkls).

The hexrd tests pass in hexrd/hexrd#293 after this change.

The exclusions were changed by loading the materials file into hexrdgui,
then modifying the reflections list, then writing out the materials file
again.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>