Fix reference data with changed/added features
Three keys were changed in the reference data: `iHKLList`, and
`planeData_excl`, `planeData_hkls`.

In the corresponding hexrd branch, the `iHKLList` in the eta omega maps
now refers to the master hkl indices rather than the reduced hkl indices.
Thus, we need to update the reference data to use that as well.

The `planeData_excl` is a newly added key, and we needed to add that here
as well.

The `planeData_hkls` now has the full hkl list and is modified by the
exclusions, rather than having a reduced hkl list.

These changes seem to fix the test.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>